[refactor] Pulling out `post` to terminal to its own file.
